I need to complete this code with the array list

+++ MyFirstCodeOOP/MyFirstCodeOOP/ContractorEmployee.cs
@@ -16,7 +16,7 @@ namespace MyFirstCodeOOP
         #region Methods
         public override decimal GetValueToPay()
         {
-            return (hours * hourvalue);
+            return ((decimal)hours * hourvalue);
         }
 
         public override string ToString()
